Comments
Problem: https://projecteuler.net/problem=74
My solution starts from the observation that the length of the chain is dictated by the first sum of digits factorials (resulted for the original number). So, first of all, I searched for the number that produce the longest chain (and this one is 1479, having the sum of digits factorials equal with 367945).
The second observation was that permutations of 1479 (and 4079 because 0! = 1) produce the longest chain.
And the last observation is that other numbers having the first chain link in permutations of 367945, also produce longest chain of 60 non-repeating terms.
The numbers (1479, 4079, 367945) are not used as given in my solution. They are calculated and used to speed the calculation of the following numbers having the same property.